size_t -> int

+++ project/parallel/batcher-odd-even.cpp
@@ -16,37 +16,37 @@ using namespace std;
 int main(int argc, char *argv[]) {
 
     //  These var names make sense when you look at the wikipedia page below
-    size_t N = 4;
+    int N = 4;
     if (argc == 2)
         N = stoi(argv[1]);
-    size_t n = pow(2, N);
+    int n = pow(2, N);
     //printf("Working with a list of 2^%i (%i)\n", N, n);
     //int ij, ijrk;
     // for ranges of loops
-    size_t rp, rk;
-    std::vector<size_t> Vnums(n);
+    int rp, rk;
+    std::vector<int> Vnums(n);
     populate_vector(Vnums, n);
 
     struct timespec start_time, end_time;
     clock_gettime(CLOCK_MONOTONIC, &start_time);
 
     // See https://en.wikipedia.org/wiki/Batcher_odd%E2%80%93even_mergesort
-    for (size_t p = 0; p < N; p++) {
+    for (int p = 0; p < N; p++) {
         rp = pow(2, p);
         //cout << 1 << std::endl;
         //#pragma omp parallel for
-        for (size_t k = 0; k < N; k++) {
+        for (int k = 0; k < N; k++) {
             //cout << 2 << std::endl;
             rk = (int)(rp / pow(2, k));
             if (rk < 1)
                 continue;
             //cout << rk << " " << k << " " << p << std::endl;
             //continue; 
-            size_t n_1_rk = n-1-rk;
+            int n_1_rk = n-1-rk;
             #pragma omp parallel for
-            for (size_t j = (rk % rp); j <= n_1_rk; j=j+2*rk) {   //  is this right?!
+            for (int j = (rk % rp); j <= n_1_rk; j=j+2*rk) {   //  is this right?!
                 //cout << omp_get_thread_num()  << std::endl;
-                for (size_t i = 0; i <= fmin(rk-1, n-j-rk-1); i++) {
+                for (int i = 0; i <= fmin(rk-1, n-j-rk-1); i++) {
                     //cout << " " << rp << " " << rk << " " << i << " " << j << " " << std::endl;
                     if (floor((i+j) / (rp*2)) == floor((i+j+rk) / (rp*2))) {
                         if ( Vnums[i+j] > Vnums[i+j+rk] ) {
